feat: add consts

+++ test.sh
@@ -7,15 +7,15 @@ SKC_EXEC="./skc"
 FAIL=0
 SUCCESS=0
 TOTAL=0
-RED="\033[31m"
-GREEN="\033[32m"
+RED="\033[91m"
+GREEN="\033[92m"
 RESET="\033[0m"
-PURPLE="\033[35m"
+PURPLE="\033[95m"
 BOLD="\033[1m"
 
 ERRORS=""
 
-echo -e $RED
+echo -e $PURPLE
 echo -e '
 ┏━  The Stańczyk Programming Language  ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
 ┃                                                                      ┃
@@ -62,27 +62,40 @@ for FILE in tests/*.sk; do
         fi
 
         ${SKC_EXEC} ${TEST_SRC}.sk -silent
-	./${TEST_SRC} > result.txt
 
-        RESULT=$(diff ${TEST_SRC}.txt result.txt)
-        echo -en $RED'┃'
-        printf "${RESET}${BOLD} ¤ %-56s${RESET}" $TEST_NAME
+        if [ "$?" == 0 ]; then
+            ./${TEST_SRC} > result.txt
 
-        if [ -z "$RESULT" ]; then
-            echo -ne "${GREEN}✔ success${RESET}"
-            SUCCESS=$(($SUCCESS + 1))
+            RESULT=$(diff ${TEST_SRC}.txt result.txt)
+            echo -en $PURPLE'┃'
+            printf "${RESET}${BOLD} ¤ %-56s${RESET}" $TEST_NAME
+
+            if [ -z "$RESULT" ]; then
+                echo -ne "${GREEN}✔ success${RESET}"
+                SUCCESS=$(($SUCCESS + 1))
+            else
+                echo -ne "${RED}✗ failed${RESET} "
+                FAIL=$(($FAIL + 1))
+                if [ "$1" == "show" ]; then
+                    ERRORS+="${BOLD}  ¤ ${TEST_NAME}.sk${RESET}\n\t- ${RESULT}\n"
+                fi
+            fi
+
+            echo -e $PURPLE'  ┃'
+            rm result.txt
+            rm ${TEST_SRC}
         else
-            echo -ne "${RED}✗ fail${RESET}   "
+            echo -en $PURPLE'┃'
+            printf "${RESET}${BOLD} ¤ %-56s${RESET}" $TEST_NAME
+
+            echo -ne "${RED}✗ failed${RESET} "
             FAIL=$(($FAIL + 1))
             if [ "$1" == "show" ]; then
                 ERRORS+="${BOLD}  ¤ ${TEST_NAME}.sk${RESET}\n\t- ${RESULT}\n"
             fi
-        fi
 
-        echo -e $RED'  ┃'
-        rm result.txt
-        rm ${TEST_SRC}
-        rm ${TEST_SRC}.c
+           echo -e $PURPLE'  ┃'
+        fi
     fi
 done
 
@@ -94,11 +107,11 @@ if [[ "$FAIL" -gt 0 ]]; then
 else
     RCOLOR="$GREEN"
 fi
-echo -e $RED'┃                                                                      ┃
+echo -e $PURPLE'┃                                                                      ┃
 ┃                                                                      ┃'
-printf "${RED}┃${RESET}                     ${RCOLOR}${BOLD}TOTAL %2d :: Success %2d | Fails %2d | Skipped %2d ${RESET}" $TOTAL $SUCCESS $FAIL $SKIPPED
-echo -e $RED'  ┃'
-echo -e $RED'┃                                                                      ┃
+printf "${PURPLE}┃${RESET}                     ${RCOLOR}${BOLD}TOTAL %2d :: Success %2d | Fails %2d | Skipped %2d ${RESET}" $TOTAL $SUCCESS $FAIL $SKIPPED
+echo -e $PURPLE'  ┃'
+echo -e $PURPLE'┃                                                                      ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━  The Stańczyk Programming Language  ━┛'
 
 if [ "$1" == "show" ]; then
